requirements.txt
fastapi
uvicorn
pandas
openpyxl
pdfplumber
python-docx
docx2txt
python-dotenv
python-multipart
gspread 
google-auth